Lookin swaggy fellas... updates below
I added color to the nav bar, changed the titles to "create account"... thought
it was a little more specific. Also I added a cancel button to second sign up
page which brings you back to homepage. changed button colors and did some other
cleaning up. It appears we are ready to move onto the the sign in interface,
then the more difficult part of the actual app interface. Cheers!