Added new gestures to the Log In and Sign Up view controllers.
Changes include hiding navigation bar when the keyboard is being used to enter text into a field. The keyboard can then be closed
with a swipe down. This swipe down will also cause the navigation bar to reappear, so that the user can move on to the next
sequence.

Changes come in each swift class under viewDidLoad and dismissKeyboardDownSwipe methods.

:D