Added backend to second sign up page
The parse cloud is now updated when the information is accurate in the four
text fields that verify emails and passwords. Next step is to locally save the
page beforehand and then use that information to update the user class on the
cloud as well. Then, user authentication for login by querying the database
to see if the user exists.