Implement card.io for new and existing users
have a working model of ability to add a card via card.io when the user
signs up or when the user is looking at his/her profile. Need to work on the
data collection and send it off to stripe so it can readily be used and
accessed on demand.